Issue Spoilage bugs fixed
