+++ Issue Spoilage/Code/Master.py
@@ -1,10 +1,7 @@
 from datetime import datetime, timedelta
 from githubAPI import GitHubAPI
 from sqlite3 import Cursor, Connection
-import Commits
-# import Issues
-# import Pulls
-import Defect_Density_Issue_Spoilage
+import Issue_Spoilage
 
 # import Lines_Of_Code_And_Num_Of_Chars
 
@@ -62,7 +59,7 @@ Calls classes and methods to analyze and interpret data.
 
         # Lines_Of_Code_And_Num_Of_Chars.Main(username, repository)
 
-        Defect_Density_Issue_Spoilage.main(self.dbCursor, self.dbConnection, datetimeList)
+        Issue_Spoilage.Main(self.dbCursor, self.dbConnection, datetimeList)
 
         # Adds all of the datetimes to the SQL database
         # Bewary of changing
