+++ Issue Spoilage/Code/sqlite_database.py
@@ -6,9 +6,9 @@ def open_connection(repo_name):
 This is some SQL code that creates the tables and columns in a database named after the repository its data is holding.
     '''    
     try:
-        connection = sqlite3.connect('/metrics/' + str(repo_name) + '_historical.db')
+        connection = sqlite3.connect('/metrics/' + str(repo_name) + '.db')
     except sqlite3.OperationalError:
-        connection = sqlite3.connect('/metrics/' + str(repo_name) + '_historical.db')
+        connection = sqlite3.connect('/metrics/' + str(repo_name) + '.db')
 
     cursor = connection.cursor()
 
@@ -90,6 +90,15 @@ This is some SQL code that creates the tables and columns in a database named af
             total_lines VARCHAR(3000),
             total_chars VARCHAR(3000));''')
 
+    cursor.execute('''CREATE TABLE IF NOT EXISTS DEFECT_DENSITY
+            (date VARCHAR(300) ,
+            DD VARCHAR(3000));''')
+
+    cursor.execute('''CREATE TABLE IF NOT EXISTS ISSUE_SPOILAGE
+            (date VARCHAR(300) ,
+            Min VARCHAR(3000),
+            Max VARCHAR(300),
+            Avg VARCHAR(300));''')
 
     connection.commit()
